requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
jsonschema>=4.20.0
pytest>=7.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
